+++ inst/NEWS.md
@@ -1,3 +1,9 @@
+# MANOVA.RM 0.4.3
+* fix ordering of data for multRM
+* update documentation
+* fix bug in simCI without interactions
+* improve error messages for missing data
+
 # MANOVA.RM 0.4.1
 * include new function multRM for multivariate repeated measures designs
 * update documentation and vignette